requirements.txt
django
django-debug-toolbar
mysqlclient
requests
psycopg2
djangorestframework
django-filter
djoser
djangorestframework-simplejwt
django-cors-headers